Added The Post Layout File(P). Set Up the CommunityFragment for implementing the ViewPager(Tabbed Transitions). Added FriendsCommunityFragment and GroupsCommunityFragment
Known Issues -
Issue with the theme of the app- Some text is not showing because it has same color as that of the background. If you can resolve, please resolve it. e.g. Tab text in CommunityFragment cant be seen because of this issue.

P.S. - Used ViewPager while implementing the Tabs. Will Update to ViewPager2 as soon as adapter is ready.

-Prajwal